Use a unified component between add and edit forms
