+++ src/routes/learn/$slug/edit.tsx
@@ -4,9 +4,11 @@ import {
   getUniqueModuleNamesFn,
   getSegmentFn,
   EditSegmentHeader,
-  SegmentForm,
+  useEditSegment,
 } from "../-components/edit-segment";
 import { Container } from "../-components/container";
+import { SegmentForm } from "../-components/segment-form";
+import { Edit } from "lucide-react";
 
 export const Route = createFileRoute("/learn/$slug/edit")({
   component: RouteComponent,
@@ -19,11 +21,31 @@ export const Route = createFileRoute("/learn/$slug/edit")({
 });
 
 function RouteComponent() {
+  const { segment, moduleNames } = Route.useLoaderData();
+  const { onSubmit, isSubmitting, uploadProgress } = useEditSegment(segment);
+
   return (
     <div className="container mx-auto">
       <EditSegmentHeader />
       <Container>
-        <SegmentForm />
+        <SegmentForm
+          headerTitle="Edit Content"
+          headerDescription="Update your course segment with rich content and media"
+          buttonText="Update Content"
+          loadingText="Updating..."
+          buttonIcon={Edit}
+          moduleNames={moduleNames}
+          onSubmit={onSubmit}
+          isSubmitting={isSubmitting}
+          uploadProgress={uploadProgress}
+          defaultValues={{
+            title: segment.title,
+            content: segment.content ?? "",
+            slug: segment.slug,
+            moduleTitle: segment.moduleTitle,
+            isPremium: segment.isPremium,
+          }}
+        />
       </Container>
     </div>
   );
