+++ src/routes/learn/-components/add-segment/index.ts
@@ -1,4 +1,3 @@
 export { AddSegmentHeader } from "./add-segment-header";
 export { useAddSegment } from "./use-add-segment";
 export { createSegmentFn, getUniqueModuleNamesFn } from "./server-functions";
-export { SegmentForm } from "./segment-form";
