+++ src/routes/learn/-components/edit-segment/index.ts
@@ -5,4 +5,3 @@ export {
   getSegmentFn,
   getUniqueModuleNamesFn,
 } from "./server-functions";
-export { SegmentForm } from "./segment-form";
