+++ src/routes/learn/-components/segment-form.tsx
@@ -14,10 +14,27 @@ import { Input } from "~/components/ui/input";
 import { Button } from "~/components/ui/button";
 import { Textarea } from "~/components/ui/textarea";
 import { Progress } from "~/components/ui/progress";
-import { Loader2 } from "lucide-react";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "~/components/ui/card";
+import { Badge } from "~/components/ui/badge";
+import {
+  Loader2,
+  FileText,
+  Link as LinkIcon,
+  Video,
+  FolderOpen,
+  Crown,
+  Upload,
+  Edit,
+  LucideIcon,
+} from "lucide-react";
 import { AutoComplete } from "~/components/ui/autocomplete";
 import { Switch } from "~/components/ui/switch";
-import { useState } from "react";
 import type { UploadProgress } from "~/utils/storage/helpers";
 
 export const formSchema = z.object({
@@ -35,187 +52,375 @@ export const formSchema = z.object({
 export type SegmentFormValues = z.infer<typeof formSchema>;
 
 interface SegmentFormProps {
-  onSubmit: (values: SegmentFormValues) => Promise<void>;
-  defaultValues: SegmentFormValues;
+  // Content customization
+  headerTitle: string;
+  headerDescription: string;
+  buttonText: string;
+  loadingText: string;
+  buttonIcon?: LucideIcon;
+
+  // Data
   moduleNames: string[];
+  defaultValues?: Partial<SegmentFormValues>;
+
+  // Functionality
+  onSubmit: (values: SegmentFormValues) => Promise<void>;
   isSubmitting: boolean;
-  submitButtonText: string;
-  submitButtonLoadingText: string;
   uploadProgress?: UploadProgress | null;
 }
 
 export function SegmentForm({
-  onSubmit,
-  defaultValues,
+  headerTitle,
+  headerDescription,
+  buttonText,
+  loadingText,
+  buttonIcon: ButtonIcon = Edit,
   moduleNames,
+  defaultValues,
+  onSubmit,
   isSubmitting,
-  submitButtonText,
-  submitButtonLoadingText,
   uploadProgress,
 }: SegmentFormProps) {
-  const form = useForm({ resolver: zodResolver(formSchema), defaultValues });
+  const form = useForm({
+    resolver: zodResolver(formSchema),
+    defaultValues: {
+      title: defaultValues?.title || "",
+      content: defaultValues?.content || "",
+      video: undefined,
+      slug: defaultValues?.slug || "",
+      moduleTitle: defaultValues?.moduleTitle || "",
+      isPremium: defaultValues?.isPremium || false,
+    },
+  });
 
   return (
-    <Form {...form}>
-      <form
-        onSubmit={form.handleSubmit(onSubmit)}
-        className="space-y-6 max-w-2xl"
-      >
-        <FormField
-          control={form.control}
-          name="title"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Title</FormLabel>
-              <FormControl>
-                <Input placeholder="Enter a title" {...field} />
-              </FormControl>
-              <FormDescription>
-                Give your content a clear and concise title.
-              </FormDescription>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
+    <div className="max-w-4xl mx-auto space-y-8">
+      {/* Header Section */}
+      <div className="text-center space-y-2">
+        <h1 className="text-3xl font-bold tracking-tight">{headerTitle}</h1>
+        <p className="text-muted-foreground text-lg">{headerDescription}</p>
+      </div>
 
-        <FormField
-          control={form.control}
-          name="slug"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Slug</FormLabel>
-              <FormControl>
-                <Input placeholder="Enter a slug" {...field} />
-              </FormControl>
-              <FormDescription>
-                The slug is used to generate the URL for your content.
-              </FormDescription>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-
-        <FormField
-          control={form.control}
-          name="content"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Content (Optional)</FormLabel>
-              <FormControl>
-                <Textarea
-                  placeholder="Enter your content"
-                  className="min-h-[200px]"
-                  {...field}
+      <Form {...form}>
+        <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
+          {/* Basic Information Card */}
+          <Card>
+            <CardHeader>
+              <CardTitle className="flex items-center gap-2">
+                <Edit className="h-5 w-5" />
+                Basic Information
+              </CardTitle>
+              <CardDescription>
+                Set the fundamental details for your content segment
+              </CardDescription>
+            </CardHeader>
+            <CardContent className="space-y-6">
+              <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+                <FormField
+                  control={form.control}
+                  name="title"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormLabel className="flex items-center gap-2">
+                        <FileText className="h-4 w-4" />
+                        Title
+                      </FormLabel>
+                      <FormControl>
+                        <Input
+                          placeholder="Enter a compelling title"
+                          className="text-base"
+                          {...field}
+                        />
+                      </FormControl>
+                      <FormDescription>
+                        Give your content a clear and engaging title
+                      </FormDescription>
+                      <FormMessage />
+                    </FormItem>
+                  )}
                 />
-              </FormControl>
-              <FormDescription>
-                Write your content using Markdown for rich formatting.
-              </FormDescription>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
 
-        <FormField
-          control={form.control}
-          name="video"
-          render={({ field: { value, onChange, ...field } }) => (
-            <FormItem>
-              <FormLabel>Video (Optional)</FormLabel>
-              <FormControl>
-                <Input
-                  type="file"
-                  accept="video/mp4"
-                  onChange={e => onChange(e.target.files?.[0])}
-                  {...field}
+                <FormField
+                  control={form.control}
+                  name="slug"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormLabel className="flex items-center gap-2">
+                        <LinkIcon className="h-4 w-4" />
+                        URL Slug
+                      </FormLabel>
+                      <FormControl>
+                        <Input
+                          placeholder="url-friendly-slug"
+                          className="text-base font-mono"
+                          {...field}
+                        />
+                      </FormControl>
+                      <FormDescription>
+                        Used to generate the URL for your content
+                      </FormDescription>
+                      <FormMessage />
+                    </FormItem>
+                  )}
                 />
-              </FormControl>
-              <FormDescription>
-                Upload a video to accompany your content.
-              </FormDescription>
-              {uploadProgress && (
-                <div className="space-y-2">
-                  <div className="flex justify-between text-sm text-muted-foreground">
-                    <span>Uploading video...</span>
-                    <span>{uploadProgress.percentage}%</span>
-                  </div>
-                  <Progress
-                    value={uploadProgress.percentage}
-                    className="w-full"
-                  />
-                  <div className="text-xs text-muted-foreground">
-                    {Math.round(uploadProgress.loaded / 1024 / 1024)} MB /{" "}
-                    {Math.round(uploadProgress.total / 1024 / 1024)} MB
-                  </div>
-                </div>
-              )}
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-
-        <FormField
-          control={form.control}
-          name="moduleTitle"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Module Name</FormLabel>
-              <FormControl>
-                <AutoComplete
-                  selectedValue={field.value}
-                  onSelectedValueChange={field.onChange}
-                  searchValue={field.value}
-                  onSearchValueChange={field.onChange}
-                  items={moduleNames.map(name => ({
-                    value: name,
-                    label: name,
-                  }))}
-                  isLoading={false}
-                  placeholder="Search or enter a module name"
-                  emptyMessage="No modules found."
-                />
-              </FormControl>
-              <FormDescription>
-                Select an existing module or enter a new one.
-              </FormDescription>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-
-        <FormField
-          control={form.control}
-          name="isPremium"
-          render={({ field }) => (
-            <FormItem className="flex flex-row items-center justify-between rounded-lg border p-4">
-              <div className="space-y-0.5">
-                <FormLabel className="text-base">Premium Content</FormLabel>
-                <FormDescription>
-                  Mark this segment as premium content. Premium content will
-                  only be accessible to premium users.
-                </FormDescription>
               </div>
-              <FormControl>
-                <Switch
-                  checked={field.value}
-                  onCheckedChange={field.onChange}
-                />
-              </FormControl>
-            </FormItem>
-          )}
-        />
 
-        <Button type="submit" disabled={isSubmitting}>
-          {isSubmitting ? (
-            <>
-              <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-              {submitButtonLoadingText}
-            </>
-          ) : (
-            submitButtonText
-          )}
-        </Button>
-      </form>
-    </Form>
+              <FormField
+                control={form.control}
+                name="moduleTitle"
+                render={({ field }) => (
+                  <FormItem>
+                    <FormLabel className="flex items-center gap-2">
+                      <FolderOpen className="h-4 w-4" />
+                      Module
+                    </FormLabel>
+                    <FormControl>
+                      <AutoComplete
+                        selectedValue={field.value}
+                        onSelectedValueChange={field.onChange}
+                        searchValue={field.value}
+                        onSearchValueChange={field.onChange}
+                        items={moduleNames.map(name => ({
+                          value: name,
+                          label: name,
+                        }))}
+                        isLoading={false}
+                        placeholder="Search or enter a module name"
+                        emptyMessage="No modules found."
+                      />
+                    </FormControl>
+                    <FormDescription>
+                      Select an existing module or create a new one
+                    </FormDescription>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+            </CardContent>
+          </Card>
+
+          {/* Content Card */}
+          <Card>
+            <CardHeader>
+              <CardTitle className="flex items-center gap-2">
+                <FileText className="h-5 w-5" />
+                Content
+              </CardTitle>
+              <CardDescription>
+                Write your lesson content using Markdown for rich formatting
+              </CardDescription>
+            </CardHeader>
+            <CardContent>
+              <FormField
+                control={form.control}
+                name="content"
+                render={({ field }) => (
+                  <FormItem>
+                    <FormLabel>Lesson Content (Optional)</FormLabel>
+                    <FormControl>
+                      <Textarea
+                        placeholder="# Welcome to the lesson&#10;&#10;Write your content here using **Markdown** formatting...&#10;&#10;- Bullet points&#10;- Code blocks&#10;- And more!"
+                        className="min-h-[300px] text-base font-mono"
+                        {...field}
+                      />
+                    </FormControl>
+                    <FormDescription>
+                      Supports Markdown syntax for headers, links, code blocks,
+                      and more
+                    </FormDescription>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+            </CardContent>
+          </Card>
+
+          {/* Media Card */}
+          <Card>
+            <CardHeader>
+              <CardTitle className="flex items-center gap-2">
+                <Video className="h-5 w-5" />
+                Media Content
+              </CardTitle>
+              <CardDescription>
+                Upload a video to enhance your lesson experience
+              </CardDescription>
+            </CardHeader>
+            <CardContent>
+              <FormField
+                control={form.control}
+                name="video"
+                render={({ field: { value, onChange, ...field } }) => (
+                  <FormItem>
+                    <FormLabel>Video File (Optional)</FormLabel>
+                    <FormControl>
+                      <div className="space-y-4">
+                        {!value ? (
+                          <div className="flex items-center justify-center w-full">
+                            <label className="flex flex-col items-center justify-center w-full h-32 border-2 border-dashed border-muted-foreground/25 rounded-lg cursor-pointer hover:bg-muted/50 transition-colors">
+                              <div className="flex flex-col items-center justify-center pt-5 pb-6">
+                                <Upload className="w-8 h-8 mb-2 text-muted-foreground" />
+                                <p className="mb-2 text-sm text-muted-foreground">
+                                  <span className="font-semibold">
+                                    Click to upload
+                                  </span>{" "}
+                                  or drag and drop
+                                </p>
+                                <p className="text-xs text-muted-foreground">
+                                  MP4 files only
+                                </p>
+                              </div>
+                              <Input
+                                type="file"
+                                accept="video/mp4"
+                                onChange={e => onChange(e.target.files?.[0])}
+                                className="hidden"
+                                {...field}
+                              />
+                            </label>
+                          </div>
+                        ) : (
+                          <Card className="bg-green-50 dark:bg-green-950/20 border-green-200 dark:border-green-800">
+                            <CardContent className="pt-6">
+                              <div className="flex items-center justify-between">
+                                <div className="flex items-center gap-3">
+                                  <div className="flex h-10 w-10 items-center justify-center rounded-full bg-green-100 dark:bg-green-900">
+                                    <Video className="h-5 w-5 text-green-600" />
+                                  </div>
+                                  <div>
+                                    <p className="text-sm font-medium">
+                                      {value.name}
+                                    </p>
+                                    <p className="text-xs text-muted-foreground">
+                                      {(value.size / 1024 / 1024).toFixed(2)} MB
+                                    </p>
+                                  </div>
+                                </div>
+                                <Button
+                                  type="button"
+                                  variant="ghost"
+                                  size="sm"
+                                  onClick={() => onChange(undefined)}
+                                  className="text-muted-foreground hover:text-destructive"
+                                >
+                                  Remove
+                                </Button>
+                              </div>
+                            </CardContent>
+                          </Card>
+                        )}
+
+                        {uploadProgress && (
+                          <Card className="bg-blue-50 dark:bg-blue-950/20 border-blue-200 dark:border-blue-800">
+                            <CardContent className="pt-6">
+                              <div className="space-y-3">
+                                <div className="flex justify-between items-center">
+                                  <div className="flex items-center gap-2">
+                                    <Loader2 className="h-4 w-4 animate-spin text-blue-600" />
+                                    <span className="text-sm font-medium">
+                                      Uploading video...
+                                    </span>
+                                  </div>
+                                  <Badge variant="secondary">
+                                    {uploadProgress.percentage}%
+                                  </Badge>
+                                </div>
+                                <Progress
+                                  value={uploadProgress.percentage}
+                                  className="w-full h-2"
+                                />
+                                <div className="flex justify-between text-xs text-muted-foreground">
+                                  <span>
+                                    {Math.round(
+                                      uploadProgress.loaded / 1024 / 1024
+                                    )}{" "}
+                                    MB uploaded
+                                  </span>
+                                  <span>
+                                    {Math.round(
+                                      uploadProgress.total / 1024 / 1024
+                                    )}{" "}
+                                    MB total
+                                  </span>
+                                </div>
+                              </div>
+                            </CardContent>
+                          </Card>
+                        )}
+                      </div>
+                    </FormControl>
+                    <FormDescription>
+                      Upload an MP4 video file to accompany your lesson content
+                    </FormDescription>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+            </CardContent>
+          </Card>
+
+          {/* Settings Card */}
+          <Card>
+            <CardHeader>
+              <CardTitle className="flex items-center gap-2">
+                <Crown className="h-5 w-5" />
+                Content Settings
+              </CardTitle>
+              <CardDescription>
+                Configure access permissions and content visibility
+              </CardDescription>
+            </CardHeader>
+            <CardContent>
+              <FormField
+                control={form.control}
+                name="isPremium"
+                render={({ field }) => (
+                  <FormItem className="flex flex-row items-center justify-between rounded-lg border p-6 bg-gradient-to-r from-yellow-50 to-amber-50 dark:from-yellow-950/20 dark:to-amber-950/20 border-yellow-200 dark:border-yellow-800">
+                    <div className="space-y-1">
+                      <FormLabel className="text-base font-semibold flex items-center gap-2">
+                        <Crown className="h-4 w-4 text-yellow-600" />
+                        Premium Content
+                      </FormLabel>
+                      <FormDescription className="text-sm">
+                        Mark this segment as premium content. Only users with
+                        premium access will be able to view this content.
+                      </FormDescription>
+                    </div>
+                    <FormControl>
+                      <Switch
+                        checked={field.value}
+                        onCheckedChange={field.onChange}
+                      />
+                    </FormControl>
+                  </FormItem>
+                )}
+              />
+            </CardContent>
+          </Card>
+
+          {/* Action Section */}
+          <div className="flex justify-center pt-4">
+            <Button
+              type="submit"
+              disabled={isSubmitting}
+              size="lg"
+              className="min-w-[200px] text-base"
+            >
+              {isSubmitting ? (
+                <>
+                  <Loader2 className="mr-2 h-5 w-5 animate-spin" />
+                  {loadingText}
+                </>
+              ) : (
+                <>
+                  <ButtonIcon className="mr-2 h-5 w-5" />
+                  {buttonText}
+                </>
+              )}
+            </Button>
+          </div>
+        </form>
+      </Form>
+    </div>
   );
 }
